Tech/s7-back/app/main.py
from contextlib import asynccontextmanager

from dotenv import load_dotenv
from starlette.responses import HTMLResponse, Response

load_dotenv()
from fastapi import FastAPI

from api import routers
from core.database import connect_to_mongo, close_mongo_connection


@asynccontextmanager
async def lifespan(app: FastAPI):
    await connect_to_mongo()
    yield
    await close_mongo_connection()


app = FastAPI(lifespan=lifespan)


for router in routers:
    app.include_router(router)



if __name__ == "__main__":
    import uvicorn
    import logging

    logging.basicConfig(level=logging.DEBUG)
    uvicorn.run(app, host="0.0.0.0", port=8000)
