Tech/s7-back/app/requirements.txt
aio-overpass==0.15.0
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.11
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ijson==3.3.0
jiter==0.9.0
lz4==4.4.3
minio==7.2.15
motor==3.7.0
multidict==6.2.0
numpy==2.2.4
openai==1.68.2
overpy==0.7
pandas==2.2.3
propcache==0.3.0
pyarrow==19.0.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
pyopensky==2.14
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rdflib==7.1.3
requests==2.32.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.39
starlette==0.46.1
tqdm==4.67.1
trino==0.333.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
Wikidata==0.8.1
yarl==1.18.3
zstandard==0.23.0
